Cambio en formato: Ahora todas las clases estan en su propio archivo java, excepto aquellas que son subclases de otras clases, estas ultimas se encuentran dentro del mismo archivo que su superclase.
Cambio en el metodo toString() de OrdenCompra: Ahora muestra los productos de la orden y sus unidades.
